Rename ancillary folder to preparation and add preparation CLI file.
