+++ pyproject.toml
@@ -24,7 +24,7 @@ Homepage = "https://github.com/GeoscienceAustralia/sar-pipleine.git"
 [project.scripts]
 find-scene = "sar_pipeline.nci.cli:find_scene_file"
 find-orbits = "sar_pipeline.nci.cli:find_orbits_for_scene"
-download-etad = "sar_pipeline.nci.cli:download_etad_for_scene"
+download-etad = "sar_pipeline.preparation.cli:download_etad"
 run-pyrosar-gamma-workflow = "sar_pipeline.nci.cli:run_pyrosar_gamma_workflow"
 submit-pyrosar-gamma-workflow = "sar_pipeline.nci.cli:submit_pyrosar_gamma_workflow"
 upload-files-in-folder-to-s3 = "sar_pipeline.nci.cli:upload_files_in_folder_to_s3"
